Added an avl-tree to search files by name
When creating a file, an existing one cannot be searched by nodeid in
the cache, as the nodeid is not yet known. So it has to be searched by
name, but rather scanning the full list of cached nodes, maintain a
second avl tree dedicated to this kind of searches.