Fixed invalidating cache when a name is unlinked or renamed
Unlinking a file implies updating its attributes and invalidating the
cached ones. As long as hard-linked files are not cached there is no
need to check other links to the same file, and only the cache entry
related to current nodeid has to be invalidated. For unlinking the
current nodeid is known, but for renaming the nodeid may have to be
fetched from user space.