Made sure not to use memcmp() with null length
When an open file is deleted, a void name is recorded so that the file
cannot be reopened. It was recorded with a length or zero which could
lead to memcmp(3) with zero length whose behavior is not clearly defined.
Define the length to be one instead (the null byte).