Fixed recording files which were deleted while open
When a file is unlinked while open, it has to be kept in the cache
until the file is closed. It is however recorded with a null name
so that the original name can be reused when creating a new file.
With this patch, the unnamed files are not indexed in the avl-tree
by name any more, because there can be several such files in a
directory, leading to a panic triggered by the duplicated key.