Fixed getting the access time updated after reads
The access time of files appeared as unchanged after reads because
the time stamps were retrieved from cache. On the first read after
a write, the data itself might be retrieved from cache, and the user
space file system was not even called, so the access time was not
updated at all. Invalidating the attribute cache in the first case
and the data cache in the second one forces the time stamps to be
retrieved from user space.