Invalidated pages fully written at end of files
Examining non-dirty pages which are not invalidated takes a long time
in pvn_vplist_dirty() for big files. Invalidating pages after they are
appended to a file inhibits this impact. However this causes some
overhead when updating some inner part of a file, which is a much less
frequent situation.