Freed private data when file creation failed
File creations are requested in two steps (mknod + open), while the
originator expect an atomic creation by open(O_CREAT). So the kernel
module merges the steps and sends a single create to user space.
When the creation fails during the second step, the private data allocated
in the first step has to be freed.