Made sure the length of file names are stored in avl trees
Usually the file known in the cache are indexed in two avl trees
(by name and by inode number). However files which are deleted while
being open are not present in the index by name. These files must be
recorded as having a name length of 0 or 1 in the index by inode number.