Double-checked whether a directory is held while unmounting
To make sure the mounted partition is not busy when unmounting, the
number of holds on the root vnode is checked. This apparently does not
catch whether the working directory of some process does not belong to
the partition, so a new check based on holds of the directories known
in cache is added.