Released the vfs hold on vnodes whose nodeid is reused
When a new file is created, the user space file system is expected to
assigns a node id not to be in use, so when the vfs has a cached
vnode with the same id, it must be an error. So free the hold on the
cached inode and log the situation. Sending a forget to the user
space is debatable, as the node id has been found ambiguous.